+++ node/resolvers/coupon.ts
@@ -1,9 +1,5 @@
 import { getNewOrderForm } from './orderForm'
 
-export const getMarketingData = (marketingData: any) => {
-  return marketingData || { coupon: '' }
-}
-
 export const mutations = {
   insertCoupon: async (_: any, args: any, ctx: Context) => {
     const {
@@ -13,6 +9,7 @@ export const mutations = {
     const newOrderForm = await checkout.insertCoupon(orderFormId!, args.text)
 
     return getNewOrderForm({
+      checkout,
       newOrderForm,
       storeGraphQL,
     })