+++ node/resolvers/index.ts
@@ -4,10 +4,20 @@ import { queries as orderFormQueries } from './orderForm'
 import { mutations as shippingMutations } from './shipping'
 
 export const resolvers = {
+  MarketingData: {
+    coupon: (marketingData: OrderFormMarketingData) => {
+      return marketingData.coupon || ''
+    },
+  },
   Mutation: {
     ...couponMutations,
     ...itemMutations,
     ...shippingMutations,
+  },
+  OrderForm: {
+    marketingData: (orderForm: OrderForm) => {
+      return orderForm.marketingData || {}
+    },
   },
   Query: {
     ...orderFormQueries,