+++ node/resolvers/orderForm.ts
@@ -1,18 +1,30 @@
+import { Checkout } from '../clients/checkout'
 import { StoreGraphQL } from '../clients/storeGraphQL'
 import { adjustItems } from './items'
+import { fillMessages } from './messages'
 import { getShippingInfo } from './shipping'
-import { getMarketingData } from './coupon'
 
 export const getNewOrderForm = async ({
+  checkout,
   newOrderForm,
   storeGraphQL,
 }: {
+  checkout: Checkout
   newOrderForm: CheckoutOrderForm
   storeGraphQL: StoreGraphQL
 }) => {
+  const { orderFormId, messages } = newOrderForm
+
+  const newMessages = fillMessages(messages)
+
+  if (messages.length) {
+    checkout.clearMessages(orderFormId)
+  }
+
   return {
     items: await adjustItems(newOrderForm.items, storeGraphQL),
-    marketingData: getMarketingData(newOrderForm.marketingData),
+    marketingData: newOrderForm.marketingData,
+    messages: newMessages,
     shipping: getShippingInfo(newOrderForm),
     totalizers: newOrderForm.totalizers,
     value: newOrderForm.value,
@@ -28,6 +40,7 @@ export const queries = {
     const newOrderForm = await checkout.orderForm()
 
     return getNewOrderForm({
+      checkout,
       newOrderForm,
       storeGraphQL,
     })